[index] convert EsIndexAdmin interface to class
Remove EsLocalIndexAdmin.
